refactor: dependencies.py 파일명에 도메인 역할 명시
