+++ app/domains/content/stage1/router.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter, Depends, HTTPException
 
 from app.common.logging.logging_config import get_logger
-from app.domains.auth.dependency.dependencies import get_current_user
+from app.domains.auth.dependency.auth_dependencies import get_current_user
 from app.domains.auth.models import User
 from app.domains.content.stage1.schemas import (
     Stage1CardsResponse,
@@ -9,7 +9,7 @@ from app.domains.content.stage1.schemas import (
     Stage1SubmitResponse,
 )
 from app.domains.content.stage1.service import stage1_pair_response
-from app.domains.progress.dependency.dependencies import get_learning_record_service
+from app.domains.progress.dependency.learning_record_dependencies import get_learning_record_service
 from app.domains.progress.service.learning_record_service import LearningRecordService
 from app.infrastructure.db.mongo.mongo_client import get_mongo_client
