+++ app/domains/content/stage3/router.py
@@ -5,7 +5,7 @@ from app.domains.content.stage3.schemas import (
     Stage3ProblemsResponse,
     Stage3ProgressResponse,
 )
-from app.domains.auth.dependency.dependencies import get_current_user
+from app.domains.auth.dependency.auth_dependencies import get_current_user
 from app.domains.auth.models import User
 from app.domains.content.stage3.service import (
     DEFAULT_STAGE3_LESSON_ID,
