+++ app/domains/instruction/controller/student_router.py
@@ -6,18 +6,18 @@ stage3_router : /student/learning/stage3       — 다음 문제 조회 / 답안
 """
 from fastapi import APIRouter, Depends, HTTPException, Query
 
-from app.domains.auth.dependency.dependencies import get_current_student
+from app.domains.auth.dependency.auth_dependencies import get_current_student
 from app.domains.auth.models import User
 from app.domains.content.stage3.schemas import Stage3AnswerRequest, Stage3AnswerResponse
 from app.domains.content.stage3.service import DEFAULT_STAGE3_LESSON_ID, get_stage3_service
-from app.domains.instruction.dependency.dependencies import get_instruction_service
+from app.domains.instruction.dependency.instruction_dependencies import get_instruction_service
 from app.domains.instruction.schema.schemas import (
     StudentAssignmentListResponse,
     StudentAssignmentProblemResponse,
     StudentAssignmentResponse,
 )
 from app.domains.instruction.service.instruction_service import InstructionService
-from app.domains.progress.dependency.dependencies import get_learning_record_service
+from app.domains.progress.dependency.learning_record_dependencies import get_learning_record_service
 from app.common.logging.logging_config import get_logger
 
 logger = get_logger(__name__)
